refactor: api 모듈을 분리한다
